+++ js/MIDIPlayer.js
@@ -1,16 +1,19 @@
 /* globals fetch, jQuery, pause, play */
 
-import { VoidPlayer } from '../js/VoidPlayer.js?v=1.7.0';
+import { VoidPlayer } from '../js/VoidPlayer.js?v=1.8.0';
 
 export class MIDIPlayer extends VoidPlayer {
   constructor (params, onUpdate, onStop, onEnd, playbackRate) {
     super(params, onUpdate, onStop, onEnd, playbackRate);
+    // TODO: playbackRate can be altered in midiPlayer by multiplying the SAMPLE_RATE,
+    //       but this distorts the pitch (which cannot be corrected in midiPlayer from what I can tell),
+    //       is tempo control at the expense of pitch worth it?
     // TODO: Use different id for different players.
     if (!jQuery('#player')[0]) {
       jQuery('body').prepend(jQuery('<div id="player">'));
       jQuery('#player').midiPlayer({
         onUpdate: onUpdate,
-        onStop: onStop,
+        onStop: onEnd,
         width: 250
       });
       jQuery('#player').hide();
@@ -77,5 +80,9 @@ export class MIDIPlayer extends VoidPlayer {
         ((prevHash << 5) - prevHash + currVal.charCodeAt(0)) | 0,
       0
     );
+  }
+
+  getTimestamp () {
+    return jQuery('#player').midiPlayer.currentTime() * 1000;
   }
 }